bug fixes in accuracyTests
Must compare dimensionless position rather than just position, since position is tracked in units of grid spacings. So, a particle at the right end of the grid will have position 100 in the coarsest grid, and ~100,000 in the finest grid. Must divide by number of grid spacings to compare. Also, when updating a grid to a new spacing, must scale the kinematic state of the particles as well, since they are held in units of [grid spacings].